% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.orthogonal.LACs.R
\name{get.orthogonal.LACs}
\alias{get.orthogonal.LACs}
\title{Generate Orthogonalized LACs}
\usage{
get.orthogonal.LACs(
  input.la,
  input.gc,
  output.la,
  id.type = c("character", "integer", "numeric"),
  select.PC = NULL,
  select.pos = NULL,
  select.id = NULL,
  if.haplo = F
)
}
\arguments{
\item{input.gc}{dataframe of global PCs of test samples. Please add sample IDs as rownames.}

\item{output.la}{filename of the output orthogonalized LACs files.}

\item{id.type}{the class of sample IDs. Possible values are "character", "integer", and "numeric".}

\item{select.PC}{a vector indicating PCs included in further analysis, e.g. c(1,2,3). Default is equivalent to the number of columns in the psi matrix.}

\item{select.pos}{a vector indicating genetic positions included in further analysis. Ascending Order. Default is set to include all genetic positions in the input.bp.}

\item{select.id}{a vector indicating sample IDs included in further analysis. Default is set to include all samples in the input.bp.}

\item{if.haplo}{a logical switch indicating whether to generate haplotype-based LACs (default = FALSE).}

\item{input.bp}{filename of original LACs files.}
}
\value{
a compressed binary file that documents sample information, genetic positions at breakpoints and orthogonalized LACs based on original LACs.
}
\description{
This function converts original LACs into orthogonalized LACs that are uncorrelated to global PCs.
}
\concept{Generating and retrieving LACs}
