% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_file_location.R
\name{get_current_file_location}
\alias{get_current_file_location}
\title{Get the location of the current file.}
\usage{
get_current_file_location()
}
\value{
The path as a string.
}
\description{
Retrieves the path of the file this function is run from. Shamelessly stolen,
and renamed, from: https://stackoverflow.com/a/55322344
}
