% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.binary.R
\name{read.binary}
\alias{read.binary}
\title{Retrieve relevant local ancestry information at breakpoints}
\usage{
read.binary(
  input.file,
  type = c("breakpoints", "LACs"),
  id.type = c("character", "integer", "numeric"),
  select.pos,
  select.id = NULL,
  if.haplo = F,
  col.names = NULL
)
}
\arguments{
\item{input.file}{filename of relevant compressed binary local ancestry files.}

\item{type}{file type of input.file. Either be "breakpoints" or "LACs".}

\item{id.type}{the class of sample IDs. Possible values are "character", "integer", and "numeric".}

\item{select.pos}{a vector indicating genetic positions included in further analysis. Ascending Order. Default is set to include all genetic positions in the input.file.}

\item{select.id}{a vector indicating sample IDs included in further analysis. Default is set to include all samples in the input.file.}

\item{if.haplo}{a logical switch indicating whether it is haplotype-based local ancestry information (default = FALSE).}

\item{col.names}{a vector indicating column names for local ancestry in local ancestry inference and PCs in LACs. For LACs files, default is set from PC1 to PCn for all columns in the input.file.}
}
\value{
a list of dataframes including genetic-position-specfic local ancestry information in either local ancestry references or in LACs.
}
\description{
This function reads and extracts compressed local ancestry information, either in local ancestry inferences or in LACs, given designated genetic positions.
}
\concept{Generating and retrieving LACs}
